+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `addPeerToAddressBook` is now async.
+
 ## [0.26.0] - 2022-09-08
 
 ### Added